move communication::fetchMessage to MessageService and use promises instead of callbacks
